refactor jointree file and split into individual implementation files
